+++ tests/ui/ui_test.py
@@ -4,10 +4,12 @@ Copyright (c) Cutleast
 
 import pytest
 
+from tests.base_test import BaseTest
+
 from ._setup.clipboard import Clipboard
 
 
-class BaseTest:
+class UiTest(BaseTest):
     """
     Base class for all ui-related tests.
     """
